perf-1968: switch from upsert to insert for bulk load (#16)
This switches from update with upsert=true to insert for the bulk load. The upsert
isn't needed as all operations are inserts. Update with upsert=true uses too much
CPU because it tries the operation as an update first.

This increases the load rate by 5X to 7X by decreasing the CPU / insert by
a factor of 5 to 7.

Simple perf results are below for MongoDB 4.2.1

Legend:
* ips - insert rate
* csec, dsec - total CPU seconds for client and mongod
* csecpi, dsecpi - CPU rate per insert (used for relative comparisons)
* secondaryIndex - yes if secondary index maintenance is also done

ips     csec    dsec    csecpi  dsecpi  secondaryIndex operation
131782  214     646     7.4     22.3    no             insert
20533   467     4046    16.1    139.6   no             upsert
94437   215     883     7.4     30.5    yes            insert
18937   458     4385    15.8    151.2   yes            upsert