Use same name for options to limit bulk insert size
The options are bulk_insert_count that is the max number of rows/docs in a bulk insert and
bulk_insert_kb that is the max number of KB for node inserts. Link and count inserts are
not limited by bulk_insert_kb because they are smaller than Node inserts.